當左側輸入區有變化才渲染內容
