+++ app/javascript/stylesheets/text_editor/chapter_list.css
@@ -1,0 +1,10 @@
+.chapter_list {
+  width: 15%;
+  height: 100vh;
+  border: lightgray 2px solid;
+  margin-right: 5px;
+  padding: 10px;
+}
+.chapter_list div{
+  margin: 5px;
+}
