+++ app/views/books/_text_editor.html.erb
@@ -1,5 +1,14 @@
-
+<h1 class="book_name"><%=  @book.title %></h1>
 <div class="wrap">
+  <div class="chapter_list">
+    <div>Chapter List</div>
+    <div>Chapter 1</div>
+    <div>session 1-1</div>
+    <div>session 1-2</div>
+    <div>Chapter 2</div>
+    <div>Chapter 3</div>
+    <div>Chapter 4</div>
+  </div>
   <div id="sourceTA" class="custom" rows="10" cols="82"></div> <%# 給 codemirror 塞 code highlight 編輯器 %>
   <div id="targetDiv"></div>                                   <%# 給 markdown-it 塞 md to html 的東西 %>
 </div>
